refacto(Results): extract campagneselector
