+++ ui/src/campagnes/ManageCampagnesPage.js
@@ -18,14 +18,7 @@ import SupportModal from "./Shared/SupportModal";
 import SuccessCreationModal from "./ManageCampagne/SuccessCreationModal";
 import { Container, ManageCampagneContainer } from "./styles/manageCampagnes.style";
 import { SearchNoResultsContainer, LoaderContainer } from "./styles/shared.style";
-import {
-  sortingKeys,
-  getChampsLibreRate,
-  getFinishedCampagnes,
-  getMedianDuration,
-  getTemoignagesCount,
-  getVerbatimsCount,
-} from "./utils";
+import { sortingKeys, getStatistics } from "./utils";
 import CampagnesTable from "./Shared/CampagnesTable/CampagnesTable";
 
 const modal = createModal({
@@ -123,14 +116,7 @@ const ManageCampagnesPage = () => {
     }
   };
 
-  const statistics = {
-    campagnesCount: campagnes?.length || 0,
-    finishedCampagnesCount: campagnes?.length ? getFinishedCampagnes(campagnes).length : 0,
-    temoignagesCount: campagnes?.length ? getTemoignagesCount(campagnes) : 0,
-    champsLibreRate: campagnes?.length ? getChampsLibreRate(campagnes) : "N/A",
-    medianDuration: campagnes?.length ? getMedianDuration(campagnes) : "N/A",
-    verbatimsCount: campagnes?.length ? getVerbatimsCount(campagnes) : "0",
-  };
+  const statistics = getStatistics(campagnes);
 
   return (
     <>
